refactor(store/sqlite): split store.go and store_test.go by topic
internal/store/sqlite/store.go (1475 lines) and store_test.go (4527
lines) were monolithic files covering all store method implementations.
Split them by entity/topic:

Production files added:
- store_domains.go    – Domain CRUD methods
- store_users.go      – User CRUD methods
- store_groups.go     – Group CRUD + set-parent methods (incl. groupGetTx,
                        groupSetParentTx transaction helpers)
- store_resources.go  – Resource CRUD methods
- store_access_types.go – AccessType CRUD methods
- store_permissions.go  – Permission CRUD methods
- store_authz.go      – Membership and permission grant/revoke methods,
                        UserAuthzResourcesList, GroupAuthzResourcesList,
                        ResourceAuthzUsersList, ResourceAuthzGroupsList,
                        PermissionMasksForUserResource, EffectiveMask,
                        and shared SQL constants/helpers

Test files added (mirrors production split):
- setup_test.go              – newTestStore helper
- store_domains_test.go      – domain store tests
- store_users_test.go        – user store tests
- store_groups_test.go       – group store tests
- store_resources_test.go    – resource store tests
- store_access_types_test.go – access-type store tests
- store_permissions_test.go  – permission store tests
- store_authz_test.go        – authz/membership store tests

store.go retains only the Store struct, New(), constraint helpers
(wrapConstraintError, constraintCode), mask helpers (maskToSQL,
maskFromSQL, SetNegativeMaskHook), and list query helpers
(sortColumns, orderByClause, likePattern, escapeLikePattern) plus
the sort-column package-level vars.

No logic changes. Coverage: 88.7% (unchanged).

Refs #101 — T65
